?Fixes for the major of issues regarding the 2.0.50 release.

+++ bin/ebuild.sh
@@ -1,14 +1,14 @@
 #!/bin/bash
 # Copyright 1999-2003 Gentoo Technologies, Inc.
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /local/data/ulm/cvs/history/var/cvsroot/gentoo-src/portage/bin/ebuild.sh,v 1.154 2004/02/04 22:40:21 carpaski Exp $
+# $Header: /local/data/ulm/cvs/history/var/cvsroot/gentoo-src/portage/bin/ebuild.sh,v 1.155 2004/02/10 01:24:56 carpaski Exp $
 
-SANDBOX_PREDICT="${SANDBOX_PREDICT}:/proc/self/maps:/dev/console"
+SANDBOX_PREDICT="${SANDBOX_PREDICT}:/proc/self/maps:/dev/console:/usr/lib/portage/pym"
 SANDBOX_WRITE="${SANDBOX_WRITE}:/dev/shm:${PORTAGE_TMPDIR}"
 SANDBOX_READ="${SANDBOX_READ}:/dev/shm:${PORTAGE_TMPDIR}"
 
 
-if [ "$*" != "depend" ] && [ "$*" != "clean" ]; then
+if [ "$*" != "depend" ] && [ "$*" != "clean" ] && [ "$*" != "nofetch" ]; then
 	if [ -f "${T}/successful" ]; then
 		rm -f "${T}/successful"
 	fi
